+++ examples/boolean_parser/main.rs
@@ -35,7 +35,7 @@ fn main() -> Result<()> {
             .into_diagnostic()
             .wrap_err(format!("Can't read file {}", file_name))?;
         let mut boolean_grammar = BooleanGrammar::new();
-        let syntax_tree = parse(&input, file_name.to_owned(), &mut boolean_grammar)
+        let syntax_tree = parse(&input, &file_name, &mut boolean_grammar)
             .wrap_err(format!("Failed parsing file {}", file_name))?;
         println!("{}", boolean_grammar);
         generate_tree_layout(&syntax_tree, &file_name).wrap_err("Error generating tree layout")