+++ examples/calc/main.rs
@@ -41,7 +41,7 @@ fn main() -> Result<()> {
             .into_diagnostic()
             .wrap_err(format!("Can't read file {}", file_name))?;
         let mut calc_grammar = CalcGrammar::new();
-        let syntax_tree = parse(&input, file_name.to_owned(), &mut calc_grammar)
+        let syntax_tree = parse(&input, &file_name, &mut calc_grammar)
             .wrap_err(format!("Failed parsing file {}", file_name))?;
         println!("{}", calc_grammar);
         generate_tree_layout(&syntax_tree, &file_name).wrap_err("Error generating tree layout")