+++ examples/keywords/main.rs
@@ -29,7 +29,7 @@ fn main() -> Result<()> {
             .into_diagnostic()
             .wrap_err(format!("Can't read file {}", file_name))?;
         let mut keywords_grammar = KeywordsGrammar::new();
-        let syntax_tree = parse(&input, file_name.to_owned(), &mut keywords_grammar)
+        let syntax_tree = parse(&input, &file_name, &mut keywords_grammar)
             .wrap_err(format!("Failed parsing file {}", file_name))?;
         println!("{}", keywords_grammar);
         generate_tree_layout(&syntax_tree, &file_name).wrap_err("Error generating tree layout")