+++ examples/scanner_states/main.rs
@@ -33,7 +33,7 @@ fn main() -> Result<()> {
             .into_diagnostic()
             .wrap_err(format!("Can't read file {}", file_name))?;
         let mut scanner_states_grammar = ScannerStatesGrammar::new();
-        let syntax_tree = parse(&input, file_name.to_owned(), &mut scanner_states_grammar)
+        let syntax_tree = parse(&input, &file_name, &mut scanner_states_grammar)
             .wrap_err(format!("Failed parsing file {}", file_name))?;
         println!("{}", scanner_states_grammar);
         generate_tree_layout(&syntax_tree, &file_name).wrap_err("Error generating tree layout")